+++ src/main/java/com/academy/communitytrackerjava/dto/ProjectDTO.java
@@ -1,10 +1,14 @@
-package com.academy.communitytrackerjava.model;
+package com.academy.communitytrackerjava.dto;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.*;
 
 @Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class ProjectDTO {
 
     private Integer projectId;
