+++ src/main/java/com/academy/communitytrackerjava/model/Project.java
@@ -1,6 +1,8 @@
 package com.academy.communitytrackerjava.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
@@ -9,6 +11,8 @@ import javax.validation.constraints.Size;
 @Entity
 @Table(name = "project")
 @Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class Project {
 
     @Id
